[ADD] GORMからDBにアクセスするプログラムを追加
