+++ main.go
@@ -5,11 +5,13 @@ import (
 	"net/http"
 	"time"
 
+	"./models"
 	"./routers"
 )
 
 func main() {
 	router := routers.InitRouter()
+	models.InitPostsModel()
 	server := &http.Server{
 		Addr:           fmt.Sprintf(":%d", 8080),
 		Handler:        router,
