+++ models/users_model.go
@@ -1,0 +1,1 @@
+package models
